Add report layout and initial version of ourdna dashboard
