--- web/src/pages/report/chart/PlotFromQuery.tsx
@@ -1,170 +1,0 @@
-import CodeIcon from '@mui/icons-material/Code'
-import OpenInFullIcon from '@mui/icons-material/OpenInFull'
-import TableChartIcon from '@mui/icons-material/TableChart'
-import {
-    Alert,
-    Box,
-    Card,
-    CardActions,
-    CardContent,
-    CircularProgress,
-    IconButton,
-    Modal,
-    Tooltip,
-    Typography,
-} from '@mui/material'
-import * as Plot from '@observablehq/plot'
-import { Table, TypeMap } from 'apache-arrow'
-import { stripIndent } from 'common-tags'
-import { ReactChild, useEffect, useRef, useState } from 'react'
-import { useMeasure } from 'react-use'
-import combineQueries from '../data/combineQueries'
-import { useProjectDbQuery } from '../data/projectDatabase'
-import { TableFromQuery } from './TableFromQuery'
-
-type PlotOptions = {
-    width: number
-}
-
-type PlotInputFunc = (
-    data: Table<TypeMap>,
-    options: PlotOptions
-) => (HTMLElement | SVGSVGElement) & Plot.Plot
-
-type QueryProps =
-    | {
-          query: string
-      }
-    | {
-          queries: { name: string; query: string }[]
-      }
-
-type Props = {
-    project: string
-    title?: ReactChild
-    subtitle?: ReactChild
-    description?: ReactChild
-    plot: PlotInputFunc
-} & QueryProps
-
-const modalStyle = {
-    position: 'absolute',
-    top: '50%',
-    left: '50%',
-    transform: 'translate(-50%, -50%)',
-    maxHeight: '100vh',
-    overflow: 'hidden',
-    width: 'calc(100% - 50px)',
-    bgcolor: 'background.paper',
-    boxShadow: 24,
-    p: 4,
-}
-
-export function PlotFromQueryCard(props: Props) {
-    const [showingTable, setShowingTable] = useState(false)
-    const [expanded, setExpanded] = useState(false)
-
-    let query = 'query' in props ? stripIndent(props.query) : combineQueries(props.queries)
-
-    return (
-        <Card sx={{ position: 'relative' }}>
-            <CardActions
-                disableSpacing
-                sx={{
-                    position: 'absolute',
-                    right: 0,
-                }}
-            >
-                <Tooltip title="View/Edit SQL" arrow>
-                    <IconButton
-                        target={'_blank'}
-                        href={`/project/${props.project}/query?query=${encodeURIComponent(query)}`}
-                    >
-                        <CodeIcon fontSize="small" />
-                    </IconButton>
-                </Tooltip>
-                <Tooltip title="View data as table" arrow>
-                    <IconButton onClick={() => setShowingTable(true)}>
-                        <TableChartIcon fontSize="small" />
-                    </IconButton>
-                </Tooltip>
-                <Tooltip title="Expand chart" arrow>
-                    <IconButton onClick={() => setExpanded(true)}>
-                        <OpenInFullIcon fontSize="small" />
-                    </IconButton>
-                </Tooltip>
-            </CardActions>
-            <CardContent>
-                <PlotFromQuery {...props} />
-            </CardContent>
-
-            <Modal open={expanded} onClose={() => setExpanded(false)}>
-                <Box sx={modalStyle}>
-                    <PlotFromQuery {...props} />
-                </Box>
-            </Modal>
-
-            <Modal open={showingTable} onClose={() => setShowingTable(false)}>
-                <Box sx={modalStyle}>
-                    <div
-                        style={{
-                            display: 'flex',
-                            maxHeight: '70vh',
-                            flexDirection: 'column',
-                        }}
-                    >
-                        <TableFromQuery query={query} project={props.project} />
-                    </div>
-                </Box>
-            </Modal>
-        </Card>
-    )
-}
-
-export function PlotFromQuery(props: Props) {
-    const containerRef = useRef<HTMLDivElement>(null)
-
-    const { project, plot } = props
-
-    const query = 'query' in props ? stripIndent(props.query) : combineQueries(props.queries)
-    const result = useProjectDbQuery(project, query)
-
-    const [measureRef, { width }] = useMeasure<HTMLDivElement>()
-
-    const data = result && result.status === 'success' ? result.data : undefined
-
-    useEffect(() => {
-        if (!data) return
-        const _plot = plot(data, { width })
-        containerRef.current?.append(_plot)
-        return () => _plot.remove()
-    }, [data, width, plot])
-
-    return (
-        <Box>
-            {(props.title || props.subtitle) && (
-                <Box mb={2}>
-                    {props.title && (
-                        <Typography fontWeight={'bold'} fontSize={16}>
-                            {props.title}
-                        </Typography>
-                    )}
-                    {props.subtitle && <Typography fontSize={14}>{props.subtitle}</Typography>}
-                </Box>
-            )}
-
-            <div ref={measureRef} style={{ width: '100%', height: 0 }} />
-            {!result || (result.status === 'loading' && <CircularProgress />)}
-            <div ref={containerRef} />
-
-            {result && result.status === 'error' && (
-                <Alert severity="error">{result.errorMessage}</Alert>
-            )}
-            {props.description && (
-                <Box mt={2}>
-                    <Typography fontSize={12}>{props.description}</Typography>
-                </Box>
-            )}
-        </Box>
-    )
-}
