+++ web/src/pages/report/components/Report.tsx
@@ -1,0 +1,9 @@
+import { Box } from '@mui/material'
+
+export default function Report(props: { children: React.ReactNode }) {
+    return (
+        <Box display={'flex'} flexDirection={'column'} gap={2}>
+            {props.children}
+        </Box>
+    )
+}
