+++ web/src/pages/report/components/ReportRow.tsx
@@ -1,0 +1,9 @@
+import { Box } from '@mui/material'
+
+export default function ReportRow(props: { children: React.ReactNode }) {
+    return (
+        <Box display={'flex'} gap={2} flexWrap={'wrap'}>
+            {props.children}
+        </Box>
+    )
+}
